Remove StreamingListenerBus and post all Streaming events to the same thread as Spark events
Including the following changes:

1. Add StreamingSparkListenerAdapter to process events in `onOtherEvent` to StreamingListener
2. Remove StreamingListenerBus
3. Merge AsynchronousListenerBus and LiveListenerBus to the same class LiveListenerBus
4. Remove unnecessary generics
5. Add `logEvent` method to SparkListenerEvent so that EventLoggingListener can use it to ignore StreamingListenerEvents
