+++ core/src/main/scala/org/apache/spark/scheduler/SparkListenerBus.scala
@@ -22,7 +22,7 @@ import org.apache.spark.util.ListenerBus
 /**
  * A [[SparkListenerEvent]] bus that relays [[SparkListenerEvent]]s to its listeners
  */
-private[spark] trait SparkListenerBus extends ListenerBus[SparkListener, SparkListenerEvent] {
+private[spark] trait SparkListenerBus extends ListenerBus {
 
   override def onPostEvent(listener: SparkListener, event: SparkListenerEvent): Unit = {
     event match {
